Started to rewrite front
